requirements.txt
Flask==3.1.2
flask-cors==6.0.1
pandas==2.3.3
h3==4.3.1
python-dotenv==1.0.1
gunicorn==22.0.0
